transfer 2 function to more compitable location
